+++ CHANGELOG.md
@@ -1,7 +1,23 @@
 # Change Log
 
 ## [Next](https://github.com/dfop02/exam-lab/tree/HEAD)
-[Full Changelog](https://github.com/dfop02/exam-lab/compare/0.1.0...HEAD)
+[Full Changelog](https://github.com/dfop02/exam-lab/compare/0.2.0...HEAD)
+
+## [0.2.0](https://github.com/dfop02/exam-lab/releases/tag/0.2.0)
+21 January 2024
+
+#### New features
+
+* Add locale for main page (support: en, pt-BR, es-ES)
+* Add basic info on main page
+* Add Review Finished Exam (review with answers)
+* Add time spent on exam at Finish Statistics
+* Update README / Add CONTRIBUTING.md with more details
+
+#### Bug fixes
+
+* Fix Review Page to fit perfectly fill all columns
+* Fix bug for not show multichoice selected options after review
 
 ## [0.1.0](https://github.com/dfop02/exam-lab/releases/tag/0.1.0)
 18 January 2024